requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
qrcode[pil]==7.4.2
Pillow==10.3.0
pydantic[email]==2.7.1
slowapi==0.1.9
python-dotenv>=1.0.0
cryptography>=42.0.0
httpx>=0.27.0
mercadopago>=2.2.0
transbank-sdk>=4.0.0
